Persisting across session / foundation for Keychain support (#6)
## Overview
This PR includes changes relating to storing private shares and signers in the Async Storage. 

I decided to put the storing logic inside the `Wallet` entity. This is a different solution that is known from other wallets, but I believe this is a reasonable approach here because, together with signers and shares, we will also introduce biometric storage. Therefore, we have a bit more data to store and we should not rely on wallets to integrate with every storage properly.

I decided to make the approach modular. I decided to create abstract classes for a different type of storage (for private keys and for signers) and immediately follow up with an opinionated implementation tailored for React Native. 
However, from the user's perspective, that is perfectly fine to replace our implementation with any other, better suited for the product. 

### Signers storage
That is the store with public information about signers connected with the given account. Those are stored in the Local Storage, being a default (although maybe suboptimal) solution for React Native Projects. 

### Private keys storage
That is the storage, which is designed to include sensitive information. Namely, we store private shares (secp256k1 shards), which should be hidden under some security measures. Temporarily, I did that also with Local Storage, which is _the wrong_ solution. However, I did it to avoid dealing with the technicalities of native code with accessing the keychain. If that PR is approved and we will have 100% confidence around every piece of the architecture, we will proceed with purely technical work of migrating those data into secure storage. 

### `CapsuleReactNativeWallet` and `CapsuleBaseWallet`.
 That is purely aesthetical change, but it's worth motivating. Since we are not opinionated toward any storage implementation and allow for replacing our implementations and predicting support for different platforms, we extracted out RN-related components. We created the `CapsuleBaseWallet` abstract class with some methods (RN-specific) implemented in the class `CapsuleReactNativeWallet`.

### `BiometricSessionManager`
Biometrics-related things were extracted out with opinionated stuff moved to `CapsuleReactNativeWallet`
 
 
 ## Testing
 In order to test it, please navigate to `contracts.ts` and uncomment  `await testSigning(newWallet)`. After opening the app for the second time, it should be possible to sign the transaction immediately (and see the output in the terminal)