Create a fake flow to create a new user (#8)
## Motivation
Now, we need to migrate from a hardcoded user-id to a generic mechanism. 

## Changes
First, note that we created a workaround to accept any code to verify a user whose email ends with `@test.usecapsule.com`. We did it to replicate the flow of creating users in the Kolektive without the tedious (and pointless) work of creating temporary react components for inputting email and verification codes. Obviously, this will need to change while making a real integration. 

Similarly, as in #6 and #7, I added an extra abstract method to be implemented in an actual wallet class and provided a default implementation for React Native which relies (again) on `AsynsStorage`. We assume that we can access the correct user-id before initiating the wallet instance. Similarly, our decision and design can be easily altered with overring methods. 

## Cookies
We needed to move around a few things to ensure that while initializing sessions manager, we already know the userId