Cookie flow (#9)
## Motivation
We want to make the complete flow work with cookies refreshing if needed.

## Changes
I decided to make a different approach than previously. 
Before, I was expecting to analyze cookies on the front end; if it's outdated, we renew the cookie and take action. Obviously, on top of that, we still had to prepare the logic to prevent the case if the cookie expired in the meantime and the server rejected us due to cookie expiration. 

The motivation for that was not to make the request that would certainly fail if we knew upfront that the cookie was expired. 
However, for the sake of MVP, we are OK with one more request, and I decided to always optimistically try to make a request, and only if the backend rejects the cookie, we make the renewal. That solution is enormously more straightforward because now we avoid all the "manual" cookie parsing, and we can safely rely on built-in React Native logic for cookie manipulation. We don't even directly store the cookie.
In the further iteration, we should implement analyzing cookies on the front end.
 
 We now have a wrapper for server fetches that automatically renews sessions if the wrapped function fails due to the cookies policy. 


~I cannot test it now due to https://capsule-labs.slack.com/archives/C042W1BSG59/p1672606731886599~

To test it, I added a 5 min delay before pre sign of tx to ensure the cookie outdated
` await new Promise((r) => setTimeout(r, 300000))`